+++ include/Intersection.h
@@ -5,8 +5,8 @@
 
 struct Intersection
 {
-    glm::vec3 point;   //The point of intersection
-    glm::vec3 normal;  //The normal of the intersected primitive on the intersection point
+    glm::dvec3 point;   //The point of intersection
+    glm::dvec3 normal;  //The normal of the intersected primitive on the intersection point
 };
 
 #endif
