+++ include/Scene.h
@@ -18,23 +18,23 @@ class Scene
 private:
   sf::Image frameBuffer;
 
-  static float Fov, RFov;
-  static float AspectRatio, ZNear, ViewportWidth, ViewportHeight;
+  static double Fov, RFov;
+  static double AspectRatio, ZNear, ViewportWidth, ViewportHeight;
 
   void GetRayFromPixel(int pixelX, int pixelY, Ray &ray);
 
   void ClearDepthBuffer();
   void ClearFrameBuffer(sf::Color clearColor);
-  float GetDepthAt(int pixelX, int pixelY);
-  void SetDepthAt(int pixelX, int pixelY, float depth);
+  double GetDepthAt(int pixelX, int pixelY);
+  void SetDepthAt(int pixelX, int pixelY, double depth);
 
 public:
   static int WindowWidth, WindowHeight;
 
   vector<Primitive*> primitives;
   vector<Light*> lights;
-  vector<float> depthBuffer;
-  float timeCount;
+  vector<double> depthBuffer;
+  double timeCount;
 
   Scene();
   virtual ~Scene();
