+++ src/Primitive.cpp
@@ -2,7 +2,7 @@
 
 Primitive::Primitive() 
 { 
-  center = glm::vec3(0.0f); 
+  center = glm::dvec3(0.0); 
 }
 
 Primitive::~Primitive() {}
