+++ src/Ray.cpp
@@ -2,11 +2,11 @@
 
 Ray::Ray()
 {
-  origin = glm::vec3(0.0f);
-  dir = glm::vec3(0.0f, 0.0f, 1.0f);
+  origin = glm::dvec3(0.0);
+  dir = glm::dvec3(0.0, 0.0, 1.0);
 }
 
-Ray::Ray(const glm::vec3 &origin,  const glm::vec3 &dir)
+Ray::Ray(const glm::dvec3 &origin,  const glm::dvec3 &dir)
 {
   this->origin = origin;
   this->dir = glm::normalize(dir);
