+++ src/Sphere.cpp
@@ -2,10 +2,10 @@
 
 Sphere::Sphere()
 {
-  radius = 1.0f;
+  radius = 1.0;
 }
 
-Sphere::Sphere(const glm::vec3& center, float radius)
+Sphere::Sphere(const glm::dvec3& center, double radius)
 {
   this->center = center;
   this->radius = radius;
@@ -15,47 +15,47 @@ Sphere::~Sphere() {}
     
 bool Sphere::GetIntersection(const Ray &ray, Intersection &intersectionResult)
 {
-    glm::vec3 rayFarPoint = ray.origin + ray.dir * 99999.9f;
+    glm::dvec3 rayFarPoint = ray.origin + ray.dir * 99999.9;
     
-    float cx = center.x, cy = center.y, cz = center.z;
-    float px = ray.origin.x, py = ray.origin.y, pz = ray.origin.z;
+    double cx = center.x, cy = center.y, cz = center.z;
+    double px = ray.origin.x, py = ray.origin.y, pz = ray.origin.z;
 
-    float vx = rayFarPoint.x - px;
-    float vy = rayFarPoint.y - py;
-    float vz = rayFarPoint.z - pz;
+    double vx = rayFarPoint.x - px;
+    double vy = rayFarPoint.y - py;
+    double vz = rayFarPoint.z - pz;
 
-    float A = vx * vx + vy * vy + vz * vz;
-    float B = 2.0f * (px * vx + py * vy + pz * vz - vx * cx - vy * cy - vz * cz);
-    float C = px * px - 2.0f * px * cx + cx * cx + py * py - 2.0f * py * cy + cy * cy +
-                pz * pz - 2.0f * pz * cz + cz * cz - radius * radius;
+    double A = vx * vx + vy * vy + vz * vz;
+    double B = 2.0 * (px * vx + py * vy + pz * vz - vx * cx - vy * cy - vz * cz);
+    double C = px * px - 2.0 * px * cx + cx * cx + py * py - 2.0 * py * cy + cy * cy +
+                pz * pz - 2.0 * pz * cz + cz * cz - radius * radius;
 
     // discriminant
-    float D = B * B - 4.0f * A * C;
+    double D = B * B - 4.0 * A * C;
 
-    if ( D < 0.0f )
+    if ( D < 0.0 )
     {
-        intersectionResult.point = glm::vec3(0.0f);
+        intersectionResult.point = glm::dvec3(0.0);
         intersectionResult.normal = glm::normalize(intersectionResult.point - center);
         return false;
     }
 
-    float t1 = ( -B - sqrt( D ) ) / ( 2.0f * A );
+    double t1 = ( -B - sqrt( D ) ) / ( 2.0 * A );
 
-    glm::vec3 solution1 = glm::vec3( ray.origin.x * ( 1.0f - t1 ) + t1 * rayFarPoint.x,
-                                     ray.origin.y * ( 1.0f - t1 ) + t1 * rayFarPoint.y,
-                                     ray.origin.z * ( 1.0f - t1 ) + t1 * rayFarPoint.z );
+    glm::dvec3 solution1 = glm::dvec3( ray.origin.x * ( 1.0 - t1 ) + t1 * rayFarPoint.x,
+                                     ray.origin.y * ( 1.0 - t1 ) + t1 * rayFarPoint.y,
+                                     ray.origin.z * ( 1.0 - t1 ) + t1 * rayFarPoint.z );
     
-    if ( D == 0.0f ) 
+    if ( D == 0.0 ) 
     { 
         intersectionResult.point = solution1;
         intersectionResult.normal = glm::normalize(intersectionResult.point - center);
         return true;
     }
     
-    float t2 = ( -B + sqrt( D ) ) / ( 2.0f * A );
-    glm::vec3 solution2 = glm::vec3( ray.origin.x * ( 1.0f - t2 ) + t2 * rayFarPoint.x,
-                                     ray.origin.y * ( 1.0f - t2 ) + t2 * rayFarPoint.y,
-                                     ray.origin.z * ( 1.0f - t2 ) + t2 * rayFarPoint.z );
+    double t2 = ( -B + sqrt( D ) ) / ( 2.0 * A );
+    glm::dvec3 solution2 = glm::dvec3( ray.origin.x * ( 1.0 - t2 ) + t2 * rayFarPoint.x,
+                                     ray.origin.y * ( 1.0 - t2 ) + t2 * rayFarPoint.y,
+                                     ray.origin.z * ( 1.0 - t2 ) + t2 * rayFarPoint.z );
 
     if (glm::length(ray.origin - solution1) < glm::length(ray.origin - solution2)) intersectionResult.point = solution1; 
     else intersectionResult.point = solution2;
